Optimized Objects.sizeof() method using new ByteCountingOutputStream.
git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/trunk/wicket@458752 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/wicket/Component.java
@@ -1095,11 +1095,11 @@ public abstract class Component implements Serializable, IBehaviorListener
 	/**
 	 * @return Size of this Component in bytes
 	 */
-	public int getSizeInBytes()
+	public long getSizeInBytes()
 	{
 		final MarkupContainer originalParent = this.parent;
 		this.parent = null;
-		final int size = Objects.sizeof(this);
+		final long size = Objects.sizeof(this);
 		this.parent = originalParent;
 		return size;
 	}
